Switches build system to Grunt and adds a "build" directory
This commit switches the build system to the Grunt task runner. This
means Node.js is required. On the plus side, there is much more
flexibiltiy in the build system without haviing to learn the arcane
language of Make.

The end result is that we have tasks like "grunt all" and "grunt pdf"
as opposed to "make all" and "make pdf".